Extract listing methods, Fix multi-rental bug
