+++ people.json
@@ -1,1 +1,1 @@
-["{\"json_class\":\"Student\",\"a\":[11,\"math\",\"ee\",\"true\",284]}"]
+["{\"json_class\":\"Student\",\"a\":[11,\"math\",\"ee\",\"true\",284]}", "{\"json_class\":\"Student\",\"a\":[22,\"math\",\"student1\",\"false\",108]}", "{\"json_class\":\"Teacher\",\"a\":[222,\"math\",\"t2\",\"true\",73]}"]
