+++ rentals.json
@@ -1,1 +1,1 @@
-[{"date":"2020-12-12","person_id":284,"book_title":"ewe"}]
+[{"date":"2020-12-12","person_id":284,"book_title":"ewe"},{"date":"2020-11-10","person_id":108,"book_title":"math1"},{"date":"2009","person_id":108,"book_title":"ewe"}]
